imageviewer improvements:  rotation by TAB/SHIFT+TAB  SHIFT with other keys reduces zoom/offset step  RGB support instead of only RGBA to save bandwidth  some other improvements for smoother UX

+++ WinPort/src/Backend/Backend.h
@@ -46,8 +46,11 @@ public:
 	virtual void OnConsoleSetCursorBlinkTime(DWORD interval) = 0;
 	virtual void OnConsoleOutputFlushDrawing() = 0;
 
+	// wgi->caps & WP_IMG_RGBA - supporting
 	virtual void OnGetConsoleImageCaps(WinportGraphicsInfo *wgi) = 0;
-	virtual bool OnSetConsoleImage(const char *id, DWORD flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
+
+	// flags: WP_IMG_*
+	virtual bool OnSetConsoleImage(const char *id, DWORD64 flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
 	virtual bool OnDeleteConsoleImage(const char *id) = 0;
 
 	virtual const char *OnConsoleBackendInfo(int entity) = 0;
@@ -225,7 +228,7 @@ public:
 	virtual void RepaintsDeferFinish(bool force) = 0;
 
 	virtual void OnGetConsoleImageCaps(WinportGraphicsInfo *wgi) = 0;
-	virtual bool OnSetConsoleImage(const char *id, DWORD flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
+	virtual bool OnSetConsoleImage(const char *id, DWORD64 flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
 	virtual bool OnDeleteConsoleImage(const char *id) = 0;
 
 	virtual const char *BackendInfo(int entity) = 0;
